+++ server/src/project/ProjectManager.ts
@@ -5,38 +5,49 @@ import {
 	TextDocuments,
 } from "vscode-languageserver";
 
-import { IAssemblerResult } from "../providers/Assembler";
 import CompletionProvider from "../providers/CompletionProvider";
 import DefinitionProvider from "../providers/DefinitionProvider";
 import DiagnosticsProvider from "../providers/DiagnosticsProvider";
+import DocumentHighlightProvider from "../providers/DocumentHighlightProvider";
 import DocumentLinkProvider from "../providers/DocumentLinkProvider";
 import DocumentSymbolProvider from "../providers/DocumentSymbolProvider";
 import HoverProvider from "../providers/HoverProvider";
+import { IPostAssemblyProvider, Provider } from "../providers/Provider";
+import ReferencesProvider from "../providers/ReferencesProvider";
+import RenameProvider from "../providers/RenameProvider";
 import SettingsProvider from "../providers/SettingsProvider";
 import { ISettings } from "../providers/SettingsProvider";
+import WorkspaceSymbolProvider from "../providers/WorkspaceSymbolProvider";
+import { IProjectInfoProvider } from "./../providers/Provider";
 import Project from "./Project";
 import { IProjectFile } from "./ProjectFiles";
 
+interface IProviderInfo {
+	provider: Provider;
+	needsPostAssemblyProcessing?: false;
+}
+
+interface IPostAssemblyProviderInfo {
+	provider: IPostAssemblyProvider;
+	needsPostAssemblyProcessing: true;
+}
+
 export default class ProjectManager {
 
-	private _connection:IConnection;
+	private readonly _connection:IConnection;
+	private readonly _providers:Array<IProviderInfo|IPostAssemblyProviderInfo>;
+	private readonly _settingsProvider:SettingsProvider;
+
 	private _projects:Project[];
 	private _currentProject?:Project;
 	private _currentDocumentUri?:string;
 	private _workspaceRoot:string;
 	private _documents:TextDocuments;
 
-	private _diagnosticsProvider:DiagnosticsProvider;
-	private _hoverProvider:HoverProvider;
-	private _definitionProvider:DefinitionProvider;
-	private _settingsProvider:SettingsProvider;
-	private _completionProvider:CompletionProvider;
-	private _documentLinkProvider:DocumentLinkProvider;
-	private _documentSymbolProvider:DocumentSymbolProvider;
-
 	constructor(connection:IConnection) {
 		this._connection = connection;
 		this._projects = [];
+
 		this._currentProject = undefined;
 		this._currentDocumentUri = undefined;
 
@@ -71,8 +82,20 @@ export default class ProjectManager {
 						resolveProvider: true,
 					},
 
+					// Symbol highlights in the document
+					documentHighlightProvider: true,
+
 					// Symbols per document
 					documentSymbolProvider: true,
+
+					// Symbol references
+					referencesProvider: true,
+
+					// Symbol renaming capabilities
+					renameProvider: true,
+
+					// Symbols per workspace
+					workspaceSymbolProvider: true,
 				},
 			};
 		});
@@ -99,21 +122,28 @@ export default class ProjectManager {
 		});
 
 		// Create providers
-		const projectInfoProvider = {
-			getEntryFiles: this.getEntryFiles.bind(this),
+		const projectInfoProvider:IProjectInfoProvider = {
+			getAllProjects: this.getAllProjects.bind(this),
+			getProjectForFile: this.getProjectForFile.bind(this),
 			getFile: this.getFile.bind(this),
-			getAssemblerResults: this.getAssemblerResults.bind(this),
 			getFileByLocalUri: this.getFileByLocalUri.bind(this),
 			getSettings: this.getSettings.bind(this),
 		};
 
-		this._diagnosticsProvider = new DiagnosticsProvider(this._connection, projectInfoProvider);
-		this._hoverProvider = new HoverProvider(this._connection, projectInfoProvider);
-		this._definitionProvider = new DefinitionProvider(this._connection, projectInfoProvider);
 		this._settingsProvider = new SettingsProvider(this._connection, projectInfoProvider);
-		this._completionProvider = new CompletionProvider(this._connection, projectInfoProvider);
-		this._documentLinkProvider = new DocumentLinkProvider(this._connection, projectInfoProvider);
-		this._documentSymbolProvider = new DocumentSymbolProvider(this._connection, projectInfoProvider);
+
+		this._providers = [
+			{ provider: new DiagnosticsProvider(this._connection, projectInfoProvider), needsPostAssemblyProcessing: true },
+			{ provider: new HoverProvider(this._connection, projectInfoProvider) },
+			{ provider: new DefinitionProvider(this._connection, projectInfoProvider) },
+			{ provider: new CompletionProvider(this._connection, projectInfoProvider) },
+			{ provider: new DocumentLinkProvider(this._connection, projectInfoProvider) },
+			{ provider: new DocumentSymbolProvider(this._connection, projectInfoProvider) },
+			{ provider: new WorkspaceSymbolProvider(this._connection, projectInfoProvider) },
+			{ provider: new DocumentHighlightProvider(this._connection, projectInfoProvider) },
+			{ provider: new ReferencesProvider(this._connection, projectInfoProvider) },
+			{ provider: new RenameProvider(this._connection, projectInfoProvider) },
+		];
 	}
 
 	public start() {
@@ -162,7 +192,7 @@ export default class ProjectManager {
 		if (document.uri !== this._currentDocumentUri) {
 			// Tab changed: change context, changing (and creating) a new project if needed
 			console.log("[pm] NEW CURRENT URI is ", document.uri);
-			const project = this.getProjectForFile(document);
+			const project = this.getProjectForDocument(document);
 
 			if (project && project !== this._currentProject) {
 				// A new project altogether, set it as default
@@ -210,24 +240,33 @@ export default class ProjectManager {
 	 * Updates existing providers with the latest info from an assembled project
 	 */
 	private updatePostAssemblyProviders(project:Project) {
-		// Diagnostics
-		this._diagnosticsProvider.process(project.getFiles(), project.getAssemblerResults());
+		for (const providerInfo of this._providers) {
+			if (providerInfo.needsPostAssemblyProcessing) {
+				providerInfo.provider.process(project.getFiles(), project.getAssemblerResults());
+			}
+		}
 	}
 
 	/**
-	 * Get a list of what is considered the "entry" file for all currently known projects
+	 * Get a list of all projects
 	 */
-	private getEntryFiles():IProjectFile[] {
-		const allEntryFiles = this._projects.map((projectInfo) => projectInfo.getEntryFileInfo());
-		return allEntryFiles.filter((file) => Boolean(file)) as IProjectFile[];
+	private getAllProjects():Project[] {
+		return this._projects.concat();
+	}
+
+	/**
+	 * Based on an URI, find the project that contains that file
+	 */
+	private getProjectForFile(uri:string):Project|undefined {
+		return this._projects.find((projectInfo) => projectInfo.hasFile(uri));
 	}
 
 	/**
 	 * Based on an URI, find a file in any of the current projects
 	 */
 	private getFile(uri:string):IProjectFile|undefined {
-		const project = this._projects.find((projectInfo) => projectInfo.hasFile(uri));
-		if (project) return project.getFileInfo(uri);
+		const project = this.getProjectForFile(uri);
+		return project ? project.getFileInfo(uri) : undefined;
 	}
 
 	/**
@@ -241,21 +280,13 @@ export default class ProjectManager {
 	}
 
 	/**
-	 * Based on the URI of any file in a project, returns its assemblying results (errors, symbols, rom, etc)
-	 */
-	private getAssemblerResults(uri:string):IAssemblerResult|undefined {
-		const project = this._projects.find((projectInfo) => projectInfo.hasFile(uri));
-		if (project) return project.getAssemblerResults();
-	}
-
-	/**
-	 * Returns the project a file belongs to
+	 * Returns the project a document belongs to
 	 * TODO: Some files (e.g. include files) can belong to more than one project!
 	 */
-	private getProjectForFile(document:TextDocument, avoidCreating:boolean = false):Project|undefined {
-		let newProject = this._projects.find((project) => project.hasFile(document.uri));
+	private getProjectForDocument(document:TextDocument, avoidCreating:boolean = false):Project|undefined {
+		let newProject = this.getProjectForFile(document.uri);
 		if (!newProject && !avoidCreating) {
-			// A new file, create a project for it
+			// A new document, create a project for it
 			newProject = new Project();
 			newProject.onAssembled.add((project) => { this.updatePostAssemblyProviders(project); });
 			newProject.addFile(document);