+++ server/src/providers/DefinitionProvider.ts
@@ -1,10 +1,11 @@
 import {
 	IConnection,
 	Location,
+	Position,
+	Range,
 	TextDocumentPositionParams,
 } from "vscode-languageserver";
 
-import { IProjectFile } from "../project/ProjectFiles";
 import LineUtils from "../utils/LineUtils";
 import { IProjectInfoProvider, Provider } from "./Provider";
 
@@ -27,35 +28,29 @@ export default class DefinitionProvider extends Provider {
 		const fileUri = textDocumentPositionParams.textDocument.uri;
 		const file = this.getProjectInfo().getFile(fileUri);
 		const sourceLines = file ? file.contentsLines : undefined;
-		const results = this.getProjectInfo().getAssemblerResults(fileUri);
+		const project = this.getProjectInfo().getProjectForFile(fileUri);
+		const results = project ? project.getAssemblerResults() : undefined;
 
-		if (sourceLines && results && !isNaN(line) && sourceLines.length > line && results.symbols) {
+		if (sourceLines && project && results && !isNaN(line) && sourceLines.length > line && results.symbols) {
 			// Find the char and the surrounding symbol it relates to
 			const sourceLine = LineUtils.removeComments(sourceLines[line]);
 			if (sourceLine) {
 				const character = textDocumentPositionParams.position.character;
-				const token = LineUtils.getTokenAtPosition(sourceLine, character);
+				const token = LineUtils.getTokenAtLinePosition(sourceLine, character);
 				const symbol = results.symbols.find((tSymbol) => tSymbol.name === token);
 				if (token && symbol && symbol.definitionLineNumber > 0) {
-					const definitionLine = symbol.definitionLineNumber - 1;
-					if (symbol.definitionFilename) {
-						// Definition is in another file
-						const otherUri:string|undefined = this.getProjectInfo().getFileByLocalUri(symbol.definitionFilename);
-						const otherFile:IProjectFile|undefined = otherUri ? this.getProjectInfo().getFile(otherUri) : undefined;
-						const otherSource:string[]|undefined = otherFile ? otherFile.contentsLines : undefined;
-
-						if (otherUri && otherSource) {
-							const tokenRange = LineUtils.getTokenRange(otherSource[definitionLine], token, definitionLine);
-							if (tokenRange) {
-								locations.push(Location.create(otherUri, tokenRange));
-							}
-						}
-					} else {
-						// Definition is in the same file
-						const tokenRange = LineUtils.getTokenRange(sourceLines[definitionLine], token, definitionLine);
-						if (tokenRange) {
-							locations.push(Location.create(textDocumentPositionParams.textDocument.uri, tokenRange));
-						}
+					const definitionFile = symbol.definitionFilename ? project.getFileInfoLocalUri(symbol.definitionFilename) : project.getEntryFileInfo();
+					if (definitionFile) {
+						const definitionLine = symbol.definitionLineNumber - 1;
+						locations.push(
+							Location.create(
+								definitionFile.uri,
+								Range.create(
+									Position.create(definitionLine, symbol.definitionColumnStart),
+									Position.create(definitionLine, symbol.definitionColumnEnd),
+								),
+							),
+						);
 					}
 				}
 			}