+++ server/src/providers/Provider.ts
@@ -1,18 +1,19 @@
 import { IConnection } from "vscode-languageserver";
 
+import Project from "../project/Project";
 import { IProjectFile } from "./../project/ProjectFiles";
 import { IAssemblerResult } from "./Assembler";
 import { ISettings } from "./SettingsProvider";
 
-export interface IPostCompilationProvider {
+export interface IPostAssemblyProvider {
 	process(files:IProjectFile[], results?:IAssemblerResult):void;
 }
 
 export interface IProjectInfoProvider {
-	getEntryFiles:() => IProjectFile[];
+	getAllProjects:() => Project[];
+	getProjectForFile:(uri:string) => Project|undefined;
 	getFile:(uri:string) => IProjectFile|undefined;
-	getAssemblerResults:(uri:string) => IAssemblerResult|undefined;
-	getFileByLocalUri:(localUri:string) => string|undefined;
+	getFileByLocalUri:(localUri:string) => IProjectFile|undefined;
 	getSettings:() => ISettings;
 }