+++ server/src/utils/StringUtils.ts
@@ -1,7 +1,3 @@
-/**
- * @author Zeh Fernando
- */
-
 export default class StringUtils {
 
 	public static splitIntoLines(text:string):string[] {